Fix newer Gtk+ build (& maybe crashes on old Gtk+)
nullToNothing allows us to wrap functions that do not return Maybe
type when they really should.  It converts the null return values
exceptions to Nothing.  It uses a type class so that the code continues
to work on newer Gtk+ where the GI info has been updated and the
function now returns Maybe.